+++ nevergrad/benchmark/experiments.py
@@ -567,7 +567,7 @@ class PackedFunctions(MultiobjectiveFunction):
         # TODO: hopefully, be able to remove it eventually
         inst = self._functions[0].instrumentation
         instf = InstrumentedFunction(PackedFunctions([f.duplicate() for f in self._functions], self._upper_bounds),
-                                     *inst.args, **inst.kwargs)
+                                     *inst.args, **inst.kwargs)   # TODO THIS FAILS
         # TODO add descriptors?
         return instf