+++ nevergrad/benchmark/xpbase.py
@@ -145,6 +145,7 @@ class Experiment:
                  cheap_constraint_checker: Optional[Callable[[Any], Any]] = None,
                  ) -> None:
         assert isinstance(function, instru.InstrumentedFunction), ("All experiment functions should derive from InstrumentedFunction")
+        assert function.dimension, "Nothing to optimize"
         self.function = function
         self.seed = seed  # depending on the inner workings of the function, the experiment may not be repeatable
         self.optimsettings = OptimizerSettings(optimizer=optimizer, num_workers=num_workers, budget=budget, batch_mode=batch_mode)